requirements.txt
crewai>=0.28.8
langchain-openai>=0.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
langchain-community
graphviz
pydantic 
pydantic-settings